+++ src/test/main.cpp
@@ -175,8 +175,7 @@ static void Test_md_sentence_count_in_bounds(lpsm::Generator& gen)
 static void Test_md_paragraph_count_equal(lpsm::Generator& gen)
 {
     std::string checking = gen.fmt_paragraph(50, lpsm::MARKDOWN);
-    int         numParas =
-            lpsm::CountParagraphs(checking, lpsm::CountParaMethod::Markdown);
+    int numParas = lpsm::CountParagraphs(checking, lpsm::Format::Markdown);
     if (numParas != 50)
     {
         LogTestFailure("md_paragraph_count_equal");
@@ -186,8 +185,7 @@ static void Test_md_paragraph_count_equal(lpsm::Generator& gen)
 static void Test_md_text_count_equal(lpsm::Generator& gen)
 {
     std::string checking = gen.fmt_text(100, lpsm::MARKDOWN);
-    int         numElems =
-            lpsm::CountParagraphs(checking, lpsm::CountParaMethod::Markdown);
+    int numElems = lpsm::CountParagraphs(checking, lpsm::Format::Markdown);
     if (numElems != 100)
     {
         LogTestFailure("md_text_count_equal");